rename sysbench folder to adhere to naming standard
The sysbench folder lacked the wrapper suffix causing it to get
snagged in pylint
