+++ snafu/sysbench_wrapper/ci_test.sh
@@ -6,6 +6,6 @@ source ci/common.sh
 
 # Build image for ci
 image_spec=$SNAFU_WRAPPER_IMAGE_PREFIX/sysbench:$SNAFU_IMAGE_TAG
-build_and_push snafu/sysbench/Dockerfile $image_spec
+build_and_push snafu/sysbench_wrapper/Dockerfile $image_spec
 pushd ripsaw
 source tests/test_sysbench.sh
